fix notification
